:bug: common_lib: modify expect code to ssh 192.168.3.100
when ssh 192.168.3.100,there have change to ask you if you want to establish
this connect, send yes/r, but sometime it will not ask.

so,I will send yes whether the ask is coming or not. if not come the yes ask,
but the password.it will fail one time.but after that , expect password is confirm
to be come.