:bug: fix bug of ge_iperf_auto_negatiation,xge_iperf_auto_negatiation
bug1: After running iperf ,it should be sleep for a while before run mii-tool

bug2: To check if auto-negatiation is excuted or not ,haifeng check last line of  txt
which generate by iperf after running mii-tool 1s.if the bandwidth is 0.00,the negatiation is running now.
but if 1s is not enough to start negatiation or negatiation is over within 1s,the check will be fail.
so the check method is change to check if there is 0.00 bandwidth in iperf txt

bug3: in xge_iperf_auto_negatiation,xge is not support auto negatiation,so mii-tool is
interrupted by fault,so no result can be grep failed. the method change to check mii-tool's return
if return is 1,the cmd is failed

other1: move iperf_killer function to common_lib to support all xge test