:bug: fix bug of xge mac tolerant
bug1: through ifconfig hw cmd ,tolerant can input a wrong value of mac,
this cmd will cause a stadard fault.haifeng use grep to get the print,but
it can not work.so I change to get $? of ifconfig to check if the mac addr
is accept or not.when sys refuse the mac addr($?=1) ,test is success.

bug2: MESSAGE=PASS is lied on the last line of function ,this cause the result
of test is always success!