:rocket: fix bug of lava stock
g_ctrlNIC var is add to select nic which to be
use as ctrl nic.but this parameter is pass only by
tester_xxx.sh.when CI call the test.this variable set as blank.
then Ci pick the ctrl nic as a test nic and change it's ip.
the nfs connect is break when ip is change!so stock is happen!

:memo: nic auto pick

:memo: fix bug of test case of hns like mtu and ipv6

:memo: add code to avoid docker exist when apt-get

:memo: add code to avoid docker exist when apt-get 1

:memo: fuck docker.io

:memo: fuck docker.io 1

:memo: fuck docker.io 2

funck

:memo: fuck docker.io 3

:memo: fuck docker.io 4

:memo: fuck docker.io 5

:memo: fuck docker.io 6

:memo: fuck docker.io 7

:memo: fix bug of setTrustRelation

:memo: fix common_lib settrustrelation bug

:memo: lava stock debug 1

:memo: lava stock debug 2

:memo: lava stock debug 3

lava stock 1

:memo: lava stock

:memo: fix ctrlNIC bug

:memo: lava stock 1

:memo: fix lava stock bug